SPACIO-30 feat: add service function
